BUG: Return decimal NaNs instead of TypeError
In case of infeasible/invalid problems npf.rate is supposed to return
nan. In cases where the required output is of type Decimal a TypeError
is raised instead as Decimal types do not support arithmatic with
floating point values. This PR uses a Decimal nan as required for the
error handling, returning a nan as expected.

See #14645