BUG: prevent underflow/overflow when finding roots in IRR
NumPy's roots function uses the companion matrix to find polynomial
roots. In the process it makes the polynomial monic by dividing by the
leading coefficient, which can cause overflow/underflow. This can be
avoided by working with a scaled version of the companion matrix
instead. Since scaling the matrix simply scales the eigenvalues
(i.e. the roots of the polynomial in this case), the original roots
can easily be recovered.

Closes gh-15.