Update _financial.py
Implemented the ability to add an initial guess to the irr function. This allows the user to guide how the final rate is chosen when more than one root is found, instead of simply choosing the value closest to zero.

This feature is related to the closed issue here: https://github.com/numpy/numpy-financial/issues/28.

For instance, the current formula returns -.76 for this cashflow: [-50, -100, 600, 300, -100]. This cashflow is quite attractive, and it is much more useful if the formula returned the other root value of 1.85. Adding the ability to have an initial guess would solve this problem.